requirements.txt
Flask
requests
beautifulsoup4
flask-cors
python-dotenv